rename all the things, more profile header styling
