--- src/components/pages/Account/AccountHeader.module.css
@@ -1,90 +1,0 @@
-.grid {
-  display: grid;
-
-  gap: var(--spacer);
-  background: var(--background-body);
-}
-
-@media (min-width: 50rem) {
-  .grid {
-    /* lazy golden ratio */
-    grid-template-columns: 1.618fr 1fr;
-  }
-
-  .description {
-    -webkit-line-clamp: 4 !important;
-  }
-}
-
-.truncate code {
-  max-width: calc(100% - 1.5rem);
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.profileInfoGrid {
-  display: grid;
-  grid-template-columns: 0.2fr 1.8fr;
-  gap: calc(var(--spacer) / 2);
-}
-
-.image {
-  width: 48px;
-  max-width: unset;
-  border-radius: 50%;
-  overflow: hidden;
-  display: inline-block;
-  margin-bottom: calc(var(--spacer) / 4);
-  border: 1px solid var(--border-color);
-  box-shadow: 0 6px 17px 0 var(--box-shadow-color);
-}
-
-.image path {
-  fill: var(--font-color-heading);
-}
-
-.infoSection {
-  border-right: 1px solid var(--border-color);
-}
-
-.name {
-  font-size: var(--font-size-title);
-  margin-bottom: calc(var(--spacer) / 8);
-}
-
-.descriptionLabel {
-  color: var(--color-secondary);
-  font-size: var(--font-size-large);
-  font-weight: var(--font-weight-bold);
-}
-
-.description {
-  color: var(--color-secondary);
-  font-size: var(--font-size-small);
-  margin-bottom: unset;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  display: -webkit-box;
-  -webkit-box-orient: vertical;
-  -webkit-line-clamp: 2;
-}
-
-.publisherLinks {
-  margin-top: calc(var(--spacer) / 2);
-}
-
-.more {
-  font-size: var(--font-size-mini);
-  margin-left: calc(var(--spacer) / 8);
-  cursor: pointer;
-}
-
-.profileInfoGrid code {
-  padding: 0;
-  color: var(--color-secondary);
-  font-size: var(--font-size-small);
-  overflow-wrap: break-word;
-  word-wrap: break-word;
-  word-break: break-word;
-}
