+++ src/components/pages/Profile/History/ComputeJobs/Details.module.css
