+++ src/components/pages/Profile/History/ComputeJobs/Results.tsx
